pyproject.toml
# pyproject.toml
# Файл конфигурации проекта

[project]
# Имя пакета, как требуется для тестов Хекслета
name = "hexlet-code"
version = "0.1.0"
description = "Task manager application built with Django"
authors = [
    { name = "Jevda", email = "alekzanya@gmail.com" }, # Замени на свои данные
]
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.2",
    "django-bootstrap5>=25.1",
    "django-filter>=25.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "whitenoise>=6.9.0",
    # Добавьте сюда Rollbar, когда дойдете до Этапа 8
    # "rollbar>=1.3.0",
]
requires-python = ">=3.10" # Указываем минимальную версию Python
readme = "README.md"
license = { text = "MIT" }

[project.urls]
Homepage = "https://github.com/Jevda/python-project-52"  # Замени на свой URL репозитория
Repository = "https://github.com/Jevda/python-project-52" # Замени на свой URL репозитория

# Секции, необходимые для сборки пакета
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# --- ИЗМЕНЕНИЯ ЗДЕСЬ ---
[tool.hatch.build.targets.wheel]
# Указываем ВСЕ пакеты (папки с __init__.py), которые должны быть включены в сборку
packages = [
  "task_manager",
  "users",
  "statuses",
  "tasks",
  "labels",
]
# --- КОНЕЦ ИЗМЕНЕНИЙ ---

[tool.ruff]
# Настройки линтера ruff (можно оставить пустыми или настроить позже)

[tool.ruff.lint]
select = ["ALL"] # Включаем все проверки ruff

[tool.ruff.format]

[dependency-groups]
dev = [
    "ruff>=0.11.7",
]
# Настройки форматирования (можно оставить пустыми или настроить позже)

# Опционально: Скрипты для запуска (если нужны)
# [project.scripts]
# my-script = "my_package.module:main_func"